+++ workalendar/tests/test_canada.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 from datetime import date
 from workalendar.tests import GenericCalendarTest
-from workalendar.canada import Canada, Ontario, Quebec, BritishColumbia
-from workalendar.canada import Alberta, Saskatchewan, Manitoba, NewBrunswick
-from workalendar.canada import NovaScotia, PrinceEdwardIsland, Newfoundland
-from workalendar.canada import Yukon, NorthwestTerritories, Nunavut
+from workalendar.america.canada import (
+    Canada, Ontario, Quebec, BritishColumbia, Alberta, Saskatchewan, Manitoba,
+    NewBrunswick, NovaScotia, PrinceEdwardIsland, Newfoundland, Yukon,
+    NorthwestTerritories, Nunavut
+)
 
 
 class CanadaTest(GenericCalendarTest):