+++ Changelog.md
@@ -4,6 +4,8 @@
 
 - Added Tests for Europe registry.
 - Added the United States of America states to the global registry (#257).
+- Added the America countries (Brazil + Mexico + Colombia + Chile + Panama + Canada) to the global registry (#257).
+- Changed Canada namespace to `workalendar.america.canada`.
 
 ## v2.6.0 (2018-08-30)